Rename autofill dropdown folder and files
